Moves requesters and events pack to the top level
